manage:test Initializing the project manager
Created the script `manage.py` from which we can run tests without
moving files around inside the project. Currently, it is only configured
to run tests (the travel time tests) but if there are ever any other
functions that we need to add (that should be directly runnable), those
should be added here.