Changed network ip grabbing.
Modified the `Network` class so that it reads this machine's ip address
instead of relying on a configuration file. Created a helper method in
`helpers.py` to make this look cleaner.

Network listening process (and conf)

Changed the `stop_listening()` method so that the process is actually
terminated. Now the `os` module sends a termination signal to the
listening process.
Changed the listening process to a daemon process so that it terminates
when it's parent process (the program) terminates. This allows for a
cleaner exit and means there won't be any zombie processes hanging
around preventing us from `bind`ing to the same ip address if we want to
run the program again.

Removed the `ip` address line from `network.conf` because we are now
reading the ip address directly from the socket.
Removed the `serial_number` line from `network.conf` because it wasn't
being used for anything.