Debugged ESC controls and added fail safe.
Debugged the ESC controller so now ESC and steering
commands can be sent between and interpreted by two Pi Cars.
Did this by putting serial communication for slave
car in one line (made it easier for Arduino to interpret it seems)
and by simplifying the Arduino slave's interpretation script.

Added a fail safe so that if the slave Arduino doesn't get a command
for 500 ms it will set its ESC and steering to neutral regardless.